+++ activerecord/test/cases/locking_test.rb
@@ -29,11 +29,9 @@ end
 class OptimisticLockingTest < ActiveRecord::TestCase
   fixtures :people, :legacy_things, :references, :string_key_objects, :peoples_treasures
 
-  def test_quote_value_passed_lock_col
+  def test_lock_version_is_incremented
     p1 = Person.find(1)
     assert_equal 0, p1.lock_version
-
-    Person.expects(:quote_value).with(0, Person.columns_hash[Person.locking_column]).returns('0').once
 
     p1.first_name = 'anika2'
     p1.save!